requirements.txt
beautifulsoup4
google-generativeai
requests
trafilatura
rich